zf - testing to see if this passes github test

+++ src/main/java/edu/ucsb/cs56/ucsb_open_lab_scheduler/controllers/TutorAssignmentController.java
@@ -64,7 +64,14 @@ public class TutorAssignmentController {
   public String manageCourse(@PathVariable("id") long id, Model model, OAuth2AuthenticationToken token,
       RedirectAttributes redirAttrs) {
     String role = authControllerAdvice.getRole(token);
-    if (!(role.equals("Admin"))) {
+    String email= (String) token.getPrincipal().getAttributes().get("email");
+    List<CourseOffering> instructorCourses=courseOfferingRepository.findByInstructorEmail(email);
+    boolean isInstructorForCourse=false;
+    for(int i=0;i<instructorCourses.size();i++){
+      if(instructorCourses.get(i).getId()==id)
+        isInstructorForCourse=true;
+    }
+    if (!((role.equals("Admin")) || (isInstructorForCourse))) {
       redirAttrs.addFlashAttribute("alertDanger", "You do not have permission to access that page");
       return "redirect:/";
     }
@@ -78,24 +85,38 @@ public class TutorAssignmentController {
     Iterable<Tutor> tutors = () -> StreamSupport.stream(tutorRepository.findAll().spliterator(), false)
         .filter(tutor -> tutor.getIsActive())
         .iterator();
-    List<TutorAssignment> tutorAssignments = tutorAssignmentRepository
-        .findByCourseOffering(courseOffering.get());
+    List<TutorAssignment> tutorAssignments = tutorAssignmentRepository.findByCourseOffering(courseOffering.get());
 
     Predicate<Tutor> shouldBeChecked = tutor -> tutorAssignments.stream()
         .filter((ta) -> ta.getTutor().getIsActive())
         .anyMatch((ta) -> ta.getTutor().equals(tutor));
+
     model.addAttribute("shouldBeChecked", shouldBeChecked);
     model.addAttribute("tutors", tutors);
     model.addAttribute("courseOffering", courseOffering.get());
+    
+    Predicate<Tutor> shouldBeChecked2 = tutor -> tutorAssignments.stream()
+      .filter((ta) -> ta.getIsCourseLead())
+      .anyMatch((ta) -> ta.getTutor().equals(tutor));
+      
+    model.addAttribute("shouldBeChecked2", shouldBeChecked2);
 
     return "tutorAssignment/manage";
   }
-
+                           
   @PostMapping("/tutorAssignment/add")
-  public ResponseEntity<?> add(@RequestParam("cid") long cid, @RequestParam("tid") long tid,
-  OAuth2AuthenticationToken token) {
+  public ResponseEntity<?> add(@RequestParam("cid") long cid, @RequestParam("tid") long tid,  @RequestParam("lead") boolean lead,
+                               OAuth2AuthenticationToken token) {
     String role = authControllerAdvice.getRole(token);
-    if (!(role.equals("Admin"))) {
+    String email= (String) token.getPrincipal().getAttributes().get("email");
+    List<CourseOffering> instructorCourses=courseOfferingRepository.findByInstructorEmail(email);
+    boolean isInstructorForCourse=false;
+    for(int i=0;i<instructorCourses.size();i++){
+      if(instructorCourses.get(i).getId()==cid)
+        isInstructorForCourse=true;
+    }
+
+    if (!((role.equals("Admin"))||(isInstructorForCourse))) {
         return new ResponseEntity<>(HttpStatus.FORBIDDEN);
     }
 
@@ -104,18 +125,57 @@ public class TutorAssignmentController {
     CourseOffering courseOffering = courseOfferingRepository.findById(cid)
         .orElseThrow(() -> new IllegalArgumentException("Invalid course offering Id:" + cid));
 
-    TutorAssignment tutorAssignment = new TutorAssignment(tutor, courseOffering);
+    TutorAssignment tutorAssignment = new TutorAssignment(tutor, courseOffering, lead);
     tutorAssignmentRepository.save(tutorAssignment);
 
     return new ResponseEntity<>(HttpStatus.OK);
   }
+    
+    @PostMapping("/tutorAssignment/update")
+    public ResponseEntity<?> update(@RequestParam("cid") long cid, @RequestParam("tid") long tid, @RequestParam("lead") boolean lead,
+                                 OAuth2AuthenticationToken token) {
+      String role = authControllerAdvice.getRole(token);
+    String email= (String) token.getPrincipal().getAttributes().get("email");
+    List<CourseOffering> instructorCourses=courseOfferingRepository.findByInstructorEmail(email);
+    boolean isInstructorForCourse=false;
+    for(int i=0;i<instructorCourses.size();i++){
+      if(instructorCourses.get(i).getId()==cid)
+        isInstructorForCourse=true;
+    }
+
+    if (!((role.equals("Admin"))||(isInstructorForCourse))) {
+        return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+    }
+
+
+
+      Tutor tutor = tutorRepository.findById(tid)
+          .orElseThrow(() -> new IllegalArgumentException("Invalid tutor Id:" + tid));
+      CourseOffering courseOffering = courseOfferingRepository.findById(cid)
+          .orElseThrow(() -> new IllegalArgumentException("Invalid course offering Id:" + cid));
+        
+        tutorAssignmentRepository.deleteByCourseOfferingIdAndTutorId(cid, tid);
+        
+        TutorAssignment tutorAssignment = new TutorAssignment(tutor, courseOffering, lead);
+        tutorAssignmentRepository.save(tutorAssignment);
+
+      return new ResponseEntity<>(HttpStatus.OK);
+    }
 
   @DeleteMapping("/tutorAssignment/{cid}/{tid}")
   public ResponseEntity<?> delete(@PathVariable("cid") long cid, @PathVariable("tid") long tid,
   OAuth2AuthenticationToken token) {
     String role = authControllerAdvice.getRole(token);
-    if (!(role.equals("Admin"))) {
-      return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+    String email= (String) token.getPrincipal().getAttributes().get("email");
+    List<CourseOffering> instructorCourses=courseOfferingRepository.findByInstructorEmail(email);
+    boolean isInstructorForCourse=false;
+    for(int i=0;i<instructorCourses.size();i++){
+      if(instructorCourses.get(i).getId()==cid)
+        isInstructorForCourse=true;
+    }
+
+    if (!((role.equals("Admin"))||(isInstructorForCourse))) {
+        return new ResponseEntity<>(HttpStatus.FORBIDDEN);
     }
 
     tutorAssignmentRepository.deleteByCourseOfferingIdAndTutorId(cid, tid);
